♻️ Refactor context cards and rename data section
