+++ static/app/types/project.tsx
@@ -59,7 +59,7 @@ export type Project = {
   builtinSymbolSources?: string[];
   defaultEnvironment?: string;
   hasUserReports?: boolean;
-  highlightContext?: string[];
+  highlightContext?: Record<string, string[]>;
   highlightTags?: string[];
   latestDeploys?: Record<string, Pick<Deploy, 'dateFinished' | 'version'>> | null;
   latestRelease?: {version: string} | null;
